+++ CHANGELOG.md
@@ -1,5 +1,17 @@
 # Changelog
 All notable changes to this project will be documented in this file.
 
+## [Unreleased] - 06.10.2021
+
+### Added
+- Added animated properties for scale and rotation to TextEffect.
+- Added option to apply effect to whole text.
+- Added Active toggle to text effect.
+
+### Changed
+- Changed DialogueTheme and DialogueSettings to ScriptableObjects.
+- Split OnLetterAppear animation curve into x and y curves.
+
+
 ## [1.0.0] - 06.10.2021
 Initial release